# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API untuk sistem pemesanan tiket bioskop.",
        "title": "Cinema Ticket API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/showtimes": {
            "get": {
                "description": "Mendapatkan semua jadwal tayang.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Showtimes"
                ],
                "summary": "Get semua jadwal tayang",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "Membuat jadwal tayang baru.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Showtimes"
                ],
                "summary": "Create jadwal tayang",
                "parameters": [
                    {
                        "description": "Create showtime request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateShowtimeRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/showtimes/{id}": {
            "get": {
                "description": "Mendapatkan detail jadwal tayang berdasarkan ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Showtimes"
                ],
                "summary": "Get jadwal tayang berdasarkan ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Showtime ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "put": {
                "description": "Mengubah data jadwal tayang.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Showtimes"
                ],
                "summary": "Update jadwal tayang",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Showtime ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update showtime request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateShowtimeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "delete": {
                "description": "Menghapus jadwal tayang.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Showtimes"
                ],
                "summary": "Delete jadwal tayang",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Showtime ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.CreateShowtimeRequest": {
            "type": "object",
            "required": [
                "cinema_id",
                "end_time",
                "movie_id",
                "price",
                "start_time",
                "studio_id"
            ],
            "properties": {
                "cinema_id": {
                    "type": "string"
                },
                "end_time": {
                    "type": "string"
                },
                "movie_id": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "start_time": {
                    "type": "string"
                },
                "studio_id": {
                    "type": "string"
                }
            }
        },
...
        "dto.UpdateShowtimeRequest": {
            "type": "object",
            "required": [
                "cinema_id",
                "end_time",
                "movie_id",
                "price",
                "start_time",
                "status",
                "studio_id"
            ],
            "properties": {
                "cinema_id": {
                    "type": "string"
                },
                "end_time": {
                    "type": "string"
                },
                "movie_id": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "start_time": {
                    "type": "string"
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "SCHEDULED",
                        "ONGOING",
                        "COMPLETED",
                        "CANCELLED"
                    ]
                },
                "studio_id": {
                    "type": "string"
                }
            }
        },
...
        "response.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    }
}